optimise download, and file reader functions
